pyproject.toml
[tool.poetry]
name = "readteam-bandit"
version = "0.1.0"
description = ""
authors = ["orla11"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
pyautogen = "^0.1.7"
docker = "^6.1.3"
black = "^23.9.1"
python-dotenv = "^1.0.0"
paramiko = "^3.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
